Implement repl inside run-show
Added rustyline dependency. This is a rust readline implementation

Made DmxOutput a trait. The original implementation is attached to
dmx_output::Live. Added an implementation that prints to stdout in
dmx_output::Stdout. Lots of refactoring around this.

Made Show::run not consume the show. Lots of refactoring around this.

Implemented repl in run-show. Had to duplicate some parsing code,
because couldn't use docopt with this easily.